Communication between components
To get the chosen query parts into one query into the input, all the steps send their data up to their parents component, which is the 'QueryMenu' component. The 'QueryMenu' component creates one query object and sends it to the 'SearchPart' component, which sends the object to his child, the 'Query' component. In the 'Query' component, the query will be shown in the input.