QueryMenu changes
Changed the descriptions of the steps. Removed the red required texts and now change the color of a step card after the user finishes a step. After finishing all steps, the "add to query" button becomes green and is able to do its job. Furthermore added the '[' and ']' to the metrics.