Changed value input and some variable names.
If the user is typing in a non numerical value the information "value should be numerical" will be shown and the border of the second step will be red. The "addtoquery" button will also not work. Until the user changes the input to a number. Furthermore added the use of the replace() function when building the query to remove all whitespaces, if there are any.