Search button functionallity, refactoring and bug fix
Added the functionallity for the search button. After clicking on search, the query string will be sent to a search function which yet doesn't exist. Send button is gray if there is no value to send and red if there is. Sending is only possible if the button is red. The condition will be changed. Maybe checking the query first, and if it makes sense the button will be red. But this will be decided in the near future. Also after clicking on the 'red' search button, the query textfield will be deleted. Furthermore did a little bit of refactoring on some places in the code.
And fixed a bug component MenuStepOne. It is necessary to set the value of the filter textbox to ' ' and call the filter1() function after clicking on the removeMetric button. Without this change the filtered metrics would still be there after clicking on removeMetric.