some changes on addQuery button, some padding adjustments and now the query stays in the textbox after searching
Changed the size of the addQuery button.  Furthermore now the query is not disappearing after starting the search, so the user can change the query in the textbox if he wishes to. Under sbt run, some ui parts where not on the places they should be. Therefore i had to add some paddings.