Changed buttons and choose metric form design. Adjusted metrics. Fixed bug and added comments
Changed the look of the "addQuery" button and the "startSearch" button.  Also changed the look of the step 1 forms. I adjusted the metric selection to the new api and added the description of metrics as a popover.
Furthermore, i fixed the bug where sometimes nothing happened after clicking on "addQuery" button. And i added a lot of comments to things that i coded.